Ignore experiment options loaded with outcomes
During the last refactor, a change was introduced that memoized experiment option lists loaded with outcomes data. These options should not probably not be included with the outcomes data in the first place, and they definitely should not replace the legal client-side options with stale or invalid server options.

In the case of #45, the server had a list `[x]` of options for an experiment and the client had a list `[x, y]`, but the server's list won.

Fixes #45